--- src/components/add/CardCompanyModal/style.module.scss
@@ -1,54 +1,0 @@
-.modal {
-    width: 375px;
-    height: 220px;
-  
-    border-radius: 5px 5px 15px 15px;
-  
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-wrap: wrap;
-  
-    background: #fff;
-    z-index: 10;
-  }
-  
-  .modal-dimmed {
-    width: 100%;
-    height: 100%;
-  
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-end;
-  
-    position: absolute;
-    top: 0;
-    left: 0;
-  
-    background: rgba(0, 0, 0, 0.5);
-  
-    border-radius: 15px;
-  
-    z-index: 5;
-  }
-  
-  .modal-item-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  .modal-item-dot {
-    margin: 0.5rem 1rem;
-    border-radius: 50%;
-    width: 2.8rem;
-    height: 2.8rem;
-    background-color: #94dacd;
-  }
-  
-  .modal-item-name {
-    font-size: 12px;
-    letter-spacing: -0.085rem;
-  }
-
