+++ src/constants/card.ts
@@ -17,3 +17,19 @@ export const PASSWORD_EACH_LENGTH = {
   MIN: 0,
   MAX: 1,
 }
+
+export const ALIAS_LENGTH = {
+  MIN: 0,
+  MAX: 10,
+}
+
+export const COMPANY_LIST = [
+  { name: '포코 카드', color: '#E24141' },
+  { name: '준 카드', color: '#547CE4' },
+  { name: '공원 카드', color: '#73BC6D' },
+  { name: '브랜 카드', color: '#DE59B9' },
+  { name: '로이드 카드', color: '#94dacd' },
+  { name: '도비 카드', color: '#E76E9A' },
+  { name: '콜린 카드', color: '#F37D3B' },
+  { name: '썬 카드', color: '#FBCD58' },
+]
